requirements.txt
opencv-contrib-python
opencv-python
uiautomator2
imutils
pillow
pytesseract
wda
baidu-aip
scikit-image